+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -330,7 +330,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KV355YM912YMYKTCMFQVMZ8Z"
+        "01KV4V0CK5S4BFN78N2VT3AMM1"
     )
     r"""The unique identifier of the data source"""