+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasetsop.py
@@ -137,7 +137,7 @@ class ListDatasetsData(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-06-14T13:29:29.915Z")
+    updated: Optional[datetime] = parse_datetime("2026-06-15T05:10:09.577Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")