+++ packages/orq-rc/src/orq_ai_sdk/models/modelmetadata.py
@@ -80,6 +80,7 @@ class ModelMetadataTypedDict(TypedDict):
     supports_b64_json_response_format: NotRequired[bool]
     supports_batch_api: NotRequired[bool]
     supports_cache_control: NotRequired[bool]
+    supports_citations: NotRequired[bool]
     supports_code_execution: NotRequired[bool]
     supports_commercial_use: NotRequired[bool]
     supports_computer_use: NotRequired[bool]
@@ -103,6 +104,7 @@ class ModelMetadataTypedDict(TypedDict):
     supports_reasoning_effort_default: NotRequired[bool]
     supports_reasoning_effort_high: NotRequired[bool]
     supports_reasoning_effort_low: NotRequired[bool]
+    supports_reasoning_effort_max: NotRequired[bool]
     supports_reasoning_effort_medium: NotRequired[bool]
     supports_reasoning_effort_minimal: NotRequired[bool]
     supports_reasoning_effort_none: NotRequired[bool]
@@ -258,6 +260,8 @@ class ModelMetadata(BaseModel):
 
     supports_cache_control: Optional[bool] = None
 
+    supports_citations: Optional[bool] = None
+
     supports_code_execution: Optional[bool] = None
 
     supports_commercial_use: Optional[bool] = None
@@ -303,6 +307,8 @@ class ModelMetadata(BaseModel):
     supports_reasoning_effort_high: Optional[bool] = None
 
     supports_reasoning_effort_low: Optional[bool] = None
+
+    supports_reasoning_effort_max: Optional[bool] = None
 
     supports_reasoning_effort_medium: Optional[bool] = None
 
@@ -419,6 +425,7 @@ class ModelMetadata(BaseModel):
                 "supports_b64_json_response_format",
                 "supports_batch_api",
                 "supports_cache_control",
+                "supports_citations",
                 "supports_code_execution",
                 "supports_commercial_use",
                 "supports_computer_use",
@@ -442,6 +449,7 @@ class ModelMetadata(BaseModel):
                 "supports_reasoning_effort_default",
                 "supports_reasoning_effort_high",
                 "supports_reasoning_effort_low",
+                "supports_reasoning_effort_max",
                 "supports_reasoning_effort_medium",
                 "supports_reasoning_effort_minimal",
                 "supports_reasoning_effort_none",