+++ packages/orq-rc/src/orq_ai_sdk/models/post_v2_feedback_evaluationop.py
@@ -55,7 +55,7 @@ class RequestBody3(BaseModel):
 
     source: Optional[PostV2FeedbackEvaluationRequestBodyFeedbackSource] = "orq"
 
-    reviewed_at: Optional[datetime] = parse_datetime("2026-06-14T13:29:47.594Z")
+    reviewed_at: Optional[datetime] = parse_datetime("2026-06-15T05:10:28.410Z")
     r"""The date and time the item was reviewed"""
 
     @model_serializer(mode="wrap")
@@ -121,7 +121,7 @@ class PostV2FeedbackEvaluationRequestBody2(BaseModel):
 
     source: Optional[PostV2FeedbackEvaluationRequestBodySource] = "orq"
 
-    reviewed_at: Optional[datetime] = parse_datetime("2026-06-14T13:29:47.594Z")
+    reviewed_at: Optional[datetime] = parse_datetime("2026-06-15T05:10:28.409Z")
     r"""The date and time the item was reviewed"""
 
     @model_serializer(mode="wrap")
@@ -187,7 +187,7 @@ class PostV2FeedbackEvaluationRequestBody1(BaseModel):
 
     source: Optional[RequestBodySource] = "orq"
 
-    reviewed_at: Optional[datetime] = parse_datetime("2026-06-14T13:29:47.593Z")
+    reviewed_at: Optional[datetime] = parse_datetime("2026-06-15T05:10:28.409Z")
     r"""The date and time the item was reviewed"""
 
     @model_serializer(mode="wrap")