+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KV355YM89RQY0V6T42Q6TM2Q"
+        "01KV4V0CK4WD50N8H4AYNHR0XK"
     )
     r"""The unique identifier of the data source"""