+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KV355YMCMNDG09SJWPQ0JEWF"
+        "01KV4V0CK8RD3VB3MJ8JNSW929"
     )
     r"""The unique identifier of the data source"""