+++ packages/orq-rc/RELEASES.md
@@ -8689,3 +8689,13 @@ Based on:
 - [python v4.11.0-rc.18] packages/orq-rc
 ### Releases
 - [PyPI v4.11.0-rc.18] https://pypi.org/project/orq-ai-sdk/4.11.0-rc.18 - packages/orq-rc
+
+## 2026-06-15 05:12:47
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.778.0 (2.904.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.11.0-rc.19] packages/orq-rc
+### Releases
+- [PyPI v4.11.0-rc.19] https://pypi.org/project/orq-ai-sdk/4.11.0-rc.19 - packages/orq-rc